Revert "break core/run cyclic dependency"
This reverts commit ed773f2e35f31a8cba0bd8ba89eadc337a662338.
