--- src/core/run/types.ts
@@ -1,26 +1,0 @@
-/*
-* types.ts
-*
-* Copyright (C) 2022 by RStudio, PBC
-*
-*/
-
-import { ProcessResult } from "../../core/process.ts";
-
-export interface RunHandlerOptions {
-  cwd?: string;
-  env?: {
-    [key: string]: string;
-  };
-  stdout?: "inherit" | "piped" | "null" | number;
-}
-
-export interface RunHandler {
-  canHandle: (script: string) => boolean;
-  run: (
-    script: string,
-    args: string[],
-    stdin?: string,
-    options?: RunHandlerOptions,
-  ) => Promise<ProcessResult>;
-}
